Fix API AdminCreateUser user creation
