+++ pkg/api/admin_users_test.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"errors"
 	"fmt"
 	"testing"
 
@@ -11,14 +12,16 @@ import (
 	"github.com/grafana/grafana/pkg/components/simplejson"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/auth"
+	"github.com/grafana/grafana/pkg/services/login"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 const (
-	testLogin        = "test@example.com"
-	testPassword     = "password"
-	nonExistingOrgID = 1000
+	testLogin         = "test@example.com"
+	testPassword      = "password"
+	nonExistingOrgID  = 1000
+	existingTestLogin = "existing@example.com"
 )
 
 func TestAdminAPIEndpoint(t *testing.T) {
@@ -223,21 +226,6 @@ func TestAdminAPIEndpoint(t *testing.T) {
 			adminCreateUserScenario(t, "Should create the user", "/api/admin/users", "/api/admin/users", createCmd, func(sc *scenarioContext) {
 				bus.ClearBusHandlers()
 
-				var userLogin string
-				var orgID int64
-
-				bus.AddHandler("test", func(cmd *models.CreateUserCommand) error {
-					userLogin = cmd.Login
-					orgID = cmd.OrgId
-
-					if orgID == nonExistingOrgID {
-						return models.ErrOrgNotFound
-					}
-
-					cmd.Result = models.User{Id: testUserID}
-					return nil
-				})
-
 				sc.fakeReqWithParams("POST", sc.url, map[string]string{}).exec()
 				assert.Equal(t, 200, sc.resp.Code)
 
@@ -245,10 +233,6 @@ func TestAdminAPIEndpoint(t *testing.T) {
 				require.NoError(t, err)
 				assert.Equal(t, testUserID, respJSON.Get("id").MustInt64())
 				assert.Equal(t, "User created", respJSON.Get("message").MustString())
-
-				// Verify that userLogin and orgID were transmitted correctly to the handler
-				assert.Equal(t, testLogin, userLogin)
-				assert.Equal(t, int64(0), orgID)
 			})
 		})
 
@@ -262,21 +246,6 @@ func TestAdminAPIEndpoint(t *testing.T) {
 			adminCreateUserScenario(t, "Should create the user", "/api/admin/users", "/api/admin/users", createCmd, func(sc *scenarioContext) {
 				bus.ClearBusHandlers()
 
-				var userLogin string
-				var orgID int64
-
-				bus.AddHandler("test", func(cmd *models.CreateUserCommand) error {
-					userLogin = cmd.Login
-					orgID = cmd.OrgId
-
-					if orgID == nonExistingOrgID {
-						return models.ErrOrgNotFound
-					}
-
-					cmd.Result = models.User{Id: testUserID}
-					return nil
-				})
-
 				sc.fakeReqWithParams("POST", sc.url, map[string]string{}).exec()
 				assert.Equal(t, 200, sc.resp.Code)
 
@@ -284,9 +253,6 @@ func TestAdminAPIEndpoint(t *testing.T) {
 				require.NoError(t, err)
 				assert.Equal(t, testUserID, respJSON.Get("id").MustInt64())
 				assert.Equal(t, "User created", respJSON.Get("message").MustString())
-
-				assert.Equal(t, testLogin, userLogin)
-				assert.Equal(t, testOrgID, orgID)
 			})
 		})
 
@@ -300,46 +266,24 @@ func TestAdminAPIEndpoint(t *testing.T) {
 			adminCreateUserScenario(t, "Should create the user", "/api/admin/users", "/api/admin/users", createCmd, func(sc *scenarioContext) {
 				bus.ClearBusHandlers()
 
-				var userLogin string
-				var orgID int64
-
-				bus.AddHandler("test", func(cmd *models.CreateUserCommand) error {
-					userLogin = cmd.Login
-					orgID = cmd.OrgId
-
-					if orgID == nonExistingOrgID {
-						return models.ErrOrgNotFound
-					}
-
-					cmd.Result = models.User{Id: testUserID}
-					return nil
-				})
-
 				sc.fakeReqWithParams("POST", sc.url, map[string]string{}).exec()
 				assert.Equal(t, 400, sc.resp.Code)
 
 				respJSON, err := simplejson.NewJson(sc.resp.Body.Bytes())
 				require.NoError(t, err)
 				assert.Equal(t, "organization not found", respJSON.Get("message").MustString())
-
-				assert.Equal(t, testLogin, userLogin)
-				assert.Equal(t, int64(1000), orgID)
 			})
 		})
 	})
 
 	t.Run("When a server admin attempts to create a user with an already existing email/login", func(t *testing.T) {
 		createCmd := dtos.AdminCreateUserForm{
-			Login:    testLogin,
+			Login:    existingTestLogin,
 			Password: testPassword,
 		}
 
 		adminCreateUserScenario(t, "Should return an error", "/api/admin/users", "/api/admin/users", createCmd, func(sc *scenarioContext) {
 			bus.ClearBusHandlers()
-
-			bus.AddHandler("test", func(cmd *models.CreateUserCommand) error {
-				return models.ErrUserAlreadyExists
-			})
 
 			sc.fakeReqWithParams("POST", sc.url, map[string]string{}).exec()
 			assert.Equal(t, 412, sc.resp.Code)
@@ -506,16 +450,43 @@ func adminCreateUserScenario(t *testing.T, desc string, url string, routePattern
 	t.Run(fmt.Sprintf("%s %s", desc, url), func(t *testing.T) {
 		t.Cleanup(bus.ClearBusHandlers)
 
+		hs := HTTPServer{
+			Bus:   bus.GetBus(),
+			Login: fakeLoginService{expected: cmd},
+		}
+
 		sc := setupScenarioContext(t, url)
 		sc.defaultHandler = routing.Wrap(func(c *models.ReqContext) response.Response {
 			sc.context = c
 			sc.context.UserId = testUserID
 
-			return AdminCreateUser(c, cmd)
+			return hs.AdminCreateUser(c, cmd)
 		})
 
 		sc.m.Post(routePattern, sc.defaultHandler)
 
 		fn(sc)
 	})
+}
+
+type fakeLoginService struct {
+	login.Service
+	expected dtos.AdminCreateUserForm
+}
+
+func (s fakeLoginService) CreateUser(cmd models.CreateUserCommand) (*models.User, error) {
+	if cmd.OrgId == nonExistingOrgID {
+		return nil, models.ErrOrgNotFound
+	}
+
+	if cmd.Login == existingTestLogin {
+		return nil, models.ErrUserAlreadyExists
+	}
+
+	if s.expected.Login == cmd.Login && s.expected.Email == cmd.Email &&
+		s.expected.Password == cmd.Password && s.expected.Name == cmd.Name && s.expected.OrgId == cmd.OrgId {
+		return &models.User{Id: testUserID}, nil
+	}
+
+	return nil, errors.New("unexpected cmd")
 }
