--- pkg/services/login/errors.go
@@ -1,9 +1,0 @@
-package login
-
-import "errors"
-
-var (
-	ErrInvalidCredentials = errors.New("invalid username or password")
-	ErrUsersQuotaReached  = errors.New("users quota reached")
-	ErrGettingUserQuota   = errors.New("error getting user quota")
-)
